fix warnings

+++ include/boost/test/impl/decorator.ipp
@@ -127,7 +127,7 @@ label::do_apply( test_unit& tu )
 void
 expected_failures::do_apply( test_unit& tu )
 {
-    tu.increase_exp_fail( m_exp_fail );
+    tu.increase_exp_fail( static_cast<unsigned int>(m_exp_fail) );
 }
 
 //____________________________________________________________________________//
